Basic card implementation is set up
I have made classes for defining cards and their movement. These interactions are modeled off of simple inventory systems and is made all in UI. I'm not sure yet if the cards should be in UI or if they would be better as sprites, but the easier implementation uses UI and event systems. I also made CardData data structures, should we want to tie specific numbers to specific icons, but this might not be necessary.